requirements-video.txt
opencv-python
numpy
pyserial
librosa
pygame
